requirements.txt
openai
python-dotenv
requests
beautifulsoup4
python-slugify
html2text

